requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2024.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0 